pyproject.toml
[tool.poetry]
name = "chat-inteligencia"
version = "0.1.0"
description = ""
authors = ["Breno"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
customtkinter = "^5.2.2"
python-dotenv = "^1.0.1"
mysql = "^0.0.3"
psycopg2 = "^2.9.10"
mysql-connector-python = "^9.1.0"
pymysql = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
